mf_core::model

Module imbl

Modules

hashmap

An unordered map.

hashset

An unordered set.

iter

Iterators over immutable data.

ordmap

An ordered map.

ordset

An ordered set.

shared_ptr

About shared pointers. Re-export the archery crate.

vector

A persistent vector.

Macros

get_in

Get a value inside multiple levels of data structures.

hashmap

Construct a hash map from a sequence of key/value pairs.

hashset

Construct a set from a sequence of values.

ordmap

Construct a map from a sequence of key/value pairs.

ordset

Construct a set from a sequence of values.

update_in

Update a value inside multiple levels of data structures.

vector

Construct a vector from a sequence of elements.

Structs

GenericHashMap

An unordered map.

GenericHashSet

An unordered set.

GenericOrdMap

An ordered map.

GenericOrdSet

An ordered set.

GenericVector

A persistent vector.

Type Aliases

HashMap

Type alias for GenericHashMap that uses std::hash::RandomState as the default hasher and DefaultSharedPtr as the pointer type.

HashSet

Type alias for GenericHashSet that uses std::hash::RandomState as the default hasher and DefaultSharedPtr as the pointer type.

OrdMap

Type alias for GenericOrdMap that uses DefaultSharedPtr as the pointer type.

OrdSet

Type alias for GenericOrdSet that uses DefaultSharedPtr as the pointer type.

Vector

Type alias for GenericVector that uses DefaultSharedPtr as the pointer type.