mf_core::model::imbl

Module hashmap

An unordered map.

An immutable hash map using hash array mapped tries.

Most operations on this map are O(logx n) for a suitably high x that it should be nearly O(1) for most maps. Because of this, it’s a great choice for a generic map as long as you don’t mind that keys will need to implement Hash and Eq.

Map entries will have a predictable order based on the hasher being used. Unless otherwise specified, this will be the standard RandomState hasher.

Structs

ConsumingIter

A consuming iterator over the elements of a map.

GenericHashMap

An unordered map.

Iter

An iterator over the elements of a map.

IterMut

A mutable iterator over the elements of a map.

Keys

An iterator over the keys of a map.

OccupiedEntry

An entry for a mapping that already exists in the map.

VacantEntry

An entry for a mapping that does not already exist in the map.

Values

An iterator over the values of a map.

Enums

Entry

A handle for a key and its associated value.

Type Aliases

HashMap

Type alias for GenericHashMap that uses std::hash::RandomState as the default hasher and DefaultSharedPtr as the pointer type.