mf_core::model::imbl

Module ordmap

An ordered map.

An immutable ordered map implemented as a [B+tree] 1.

Most operations on this type of map are O(log n). A HashMap is usually a better choice for performance, but the OrdMap has the advantage of only requiring an Ord constraint on the key, and of being ordered, so that keys always come out from lowest to highest, where a HashMap has no guaranteed ordering.

Structs

ConsumingIter

A consuming iterator over the elements of a map.

DiffIter

An iterator over the differences between two maps.

GenericOrdMap

An ordered map.

Iter

An iterator over the key/value pairs of a map.

Keys

An iterator ove the keys of a map.

OccupiedEntry

An entry for a mapping that already exists in the map.

RangedIter

An iterator over a range of key/value pairs in a map.

VacantEntry

An entry for a mapping that does not already exist in the map.

Values

An iterator over the values of a map.

Enums

DiffItem

A description of a difference between two ordered maps.

Entry

A handle for a key and its associated value.

Type Aliases

OrdMap

Type alias for GenericOrdMap that uses DefaultSharedPtr as the pointer type.