mf_core::model::imbl::ordmap

Struct GenericOrdMap

pub struct GenericOrdMap<K, V, P>
where
    P: SharedPointerKind,
{ /* private fields */ }

An ordered map.

An immutable ordered map implemented as a B+tree 1.

Most operations on this type of map are O(log n). A HashMap is usually a better choice for performance, but the OrdMap has the advantage of only requiring an Ord constraint on the key, and of being ordered, so that keys always come out from lowest to highest, where a HashMap has no guaranteed ordering.

Implementations

impl<K, V, P> GenericOrdMap<K, V, P>
where P: SharedPointerKind,

pub fn new() -> GenericOrdMap<K, V, P>

Construct an empty map.

pub fn unit(key: K, value: V) -> GenericOrdMap<K, V, P>

Construct a map with a single mapping.

Examples

let map = OrdMap::unit(123, "onetwothree");
assert_eq!(
  map.get(&123),
  Some(&"onetwothree")
);

pub fn is_empty(&self) -> bool

Test whether a map is empty.

Time: O(1)

Examples

assert!(
  !ordmap!{1 => 2}.is_empty()
);
assert!(
  OrdMap::<i32, i32>::new().is_empty()
);

pub fn ptr_eq(&self, other: &GenericOrdMap<K, V, P>) -> bool

Test whether two maps refer to the same content in memory.

This is true if the two sides are references to the same map, or if the two maps refer to the same root node.

This would return true if you’re comparing a map to itself, or if you’re comparing a map to a fresh clone of itself.

Time: O(1)

pub fn len(&self) -> usize

Get the size of a map.

Time: O(1)

Examples

assert_eq!(3, ordmap!{
  1 => 11,
  2 => 22,
  3 => 33
}.len());

pub fn clear(&mut self)

Discard all elements from the map.

This leaves you with an empty map, and all elements that were previously inside it are dropped.

Time: O(n)

Examples

let mut map = ordmap![1=>1, 2=>2, 3=>3];
map.clear();
assert!(map.is_empty());

impl<K, V, P> GenericOrdMap<K, V, P>
where K: Ord, P: SharedPointerKind,

pub fn get_max(&self) -> Option<&(K, V)>

Get the largest key in a map, along with its value. If the map is empty, return None.

Time: O(log n)

Examples

assert_eq!(Some(&(3, 33)), ordmap!{
  1 => 11,
  2 => 22,
  3 => 33
}.get_max());

pub fn get_min(&self) -> Option<&(K, V)>

Get the smallest key in a map, along with its value. If the map is empty, return None.

Time: O(log n)

Examples

assert_eq!(Some(&(1, 11)), ordmap!{
  1 => 11,
  2 => 22,
  3 => 33
}.get_min());

pub fn iter(&self) -> Iter<'_, K, V, P>

Get an iterator over the key/value pairs of a map.

pub fn range<R, BK>(&self, range: R) -> RangedIter<'_, K, V, P>
where R: RangeBounds<BK>, K: Borrow<BK>, BK: Ord + ?Sized,

Create an iterator over a range of key/value pairs.

pub fn keys(&self) -> Keys<'_, K, V, P>

Get an iterator over a map’s keys.

pub fn values(&self) -> Values<'_, K, V, P>

Get an iterator over a map’s values.

pub fn diff<'a, 'b>( &'a self, other: &'b GenericOrdMap<K, V, P>, ) -> DiffIter<'a, 'b, K, V, P>

Get an iterator over the differences between this map and another, i.e. the set of entries to add, update, or remove to this map in order to make it equal to the other map.

This function will avoid visiting nodes which are shared between the two sets, meaning that even very large sets can be compared quickly if most of their structure is shared.

Time: O(n) where n is the size of the larger map.

pub fn get<BK>(&self, key: &BK) -> Option<&V>
where BK: Ord + ?Sized, K: Borrow<BK>,

Get the value for a key from a map.

Time: O(log n)

Examples

let map = ordmap!{123 => "lol"};
assert_eq!(
  map.get(&123),
  Some(&"lol")
);

pub fn get_key_value<BK>(&self, key: &BK) -> Option<(&K, &V)>
where BK: Ord + ?Sized, K: Borrow<BK>,

Get the key/value pair for a key from a map.

Time: O(log n)

Examples

let map = ordmap!{123 => "lol"};
assert_eq!(
  map.get_key_value(&123),
  Some((&123, &"lol"))
);

pub fn get_prev<BK>(&self, key: &BK) -> Option<(&K, &V)>
where BK: Ord + ?Sized, K: Borrow<BK>,

Get a reference to the closest smaller entry in a map to a given key.

If the map contains the given key, this is returned. Otherwise, the closest key in the map smaller than the given value is returned. If the smallest key in the map is larger than the given key, None is returned.

Examples

let map = ordmap![1 => 1, 3 => 3, 5 => 5];
assert_eq!(Some((&3, &3)), map.get_prev(&4));

pub fn get_next<BK>(&self, key: &BK) -> Option<(&K, &V)>
where BK: Ord + ?Sized, K: Borrow<BK>,

Get a reference to the closest larger entry in a map to a given key.

If the set contains the given value, this is returned. Otherwise, the closest value in the set larger than the given value is returned. If the largest value in the set is smaller than the given value, None is returned.

Examples

let map = ordmap![1 => 1, 3 => 3, 5 => 5];
assert_eq!(Some((&5, &5)), map.get_next(&4));

pub fn contains_key<BK>(&self, k: &BK) -> bool
where BK: Ord + ?Sized, K: Borrow<BK>,

Test for the presence of a key in a map.

Time: O(log n)

Examples

let map = ordmap!{123 => "lol"};
assert!(
  map.contains_key(&123)
);
assert!(
  !map.contains_key(&321)
);

pub fn is_submap_by<B, RM, F, P2>(&self, other: RM, cmp: F) -> bool
where F: FnMut(&V, &B) -> bool, RM: Borrow<GenericOrdMap<K, B, P2>>, P2: SharedPointerKind,

Test whether a map is a submap of another map, meaning that all keys in our map must also be in the other map, with the same values.

Use the provided function to decide whether values are equal.

Time: O(n log n)

pub fn is_proper_submap_by<B, RM, F, P2>(&self, other: RM, cmp: F) -> bool
where F: FnMut(&V, &B) -> bool, RM: Borrow<GenericOrdMap<K, B, P2>>, P2: SharedPointerKind,

Test whether a map is a proper submap of another map, meaning that all keys in our map must also be in the other map, with the same values. To be a proper submap, ours must also contain fewer keys than the other map.

Use the provided function to decide whether values are equal.

Time: O(n log n)

pub fn is_submap<RM>(&self, other: RM) -> bool
where V: PartialEq, RM: Borrow<GenericOrdMap<K, V, P>>,

Test whether a map is a submap of another map, meaning that all keys in our map must also be in the other map, with the same values.

Time: O(n log n)

Examples

let map1 = ordmap!{1 => 1, 2 => 2};
let map2 = ordmap!{1 => 1, 2 => 2, 3 => 3};
assert!(map1.is_submap(map2));

pub fn is_proper_submap<RM>(&self, other: RM) -> bool
where V: PartialEq, RM: Borrow<GenericOrdMap<K, V, P>>,

Test whether a map is a proper submap of another map, meaning that all keys in our map must also be in the other map, with the same values. To be a proper submap, ours must also contain fewer keys than the other map.

Time: O(n log n)

Examples

let map1 = ordmap!{1 => 1, 2 => 2};
let map2 = ordmap!{1 => 1, 2 => 2, 3 => 3};
assert!(map1.is_proper_submap(map2));

let map3 = ordmap!{1 => 1, 2 => 2};
let map4 = ordmap!{1 => 1, 2 => 2};
assert!(!map3.is_proper_submap(map4));

impl<K, V, P> GenericOrdMap<K, V, P>
where K: Ord + Clone, V: Clone, P: SharedPointerKind,

pub fn get_mut<BK>(&mut self, key: &BK) -> Option<&mut V>
where BK: Ord + ?Sized, K: Borrow<BK>,

Get a mutable reference to the value for a key from a map.

Time: O(log n)

Examples

let mut map = ordmap!{123 => "lol"};
if let Some(value) = map.get_mut(&123) {
    *value = "omg";
}
assert_eq!(
  map.get(&123),
  Some(&"omg")
);

pub fn get_key_value_mut<BK>(&mut self, key: &BK) -> Option<(&K, &mut V)>
where BK: Ord + ?Sized, K: Borrow<BK>,

Get the key/value pair for a key from a map.

Time: O(log n)

Examples

let mut map = ordmap!{123 => "lol"};
assert_eq!(
  map.get_key_value_mut(&123),
  Some((&123, &mut "lol"))
);

pub fn get_prev_mut<BK>(&mut self, key: &BK) -> Option<(&K, &mut V)>
where BK: Ord + ?Sized, K: Borrow<BK>,

Get the closest smaller entry in a map to a given key as a mutable reference.

If the map contains the given key, this is returned. Otherwise, the closest key in the map smaller than the given value is returned. If the smallest key in the map is larger than the given key, None is returned.

Examples

let mut map = ordmap![1 => 1, 3 => 3, 5 => 5];
if let Some((key, value)) = map.get_prev_mut(&4) {
    *value = 4;
}
assert_eq!(ordmap![1 => 1, 3 => 4, 5 => 5], map);

pub fn get_next_mut<BK>(&mut self, key: &BK) -> Option<(&K, &mut V)>
where BK: Ord + ?Sized, K: Borrow<BK>,

Get the closest larger entry in a map to a given key as a mutable reference.

If the set contains the given value, this is returned. Otherwise, the closest value in the set larger than the given value is returned. If the largest value in the set is smaller than the given value, None is returned.

Examples

let mut map = ordmap![1 => 1, 3 => 3, 5 => 5];
if let Some((key, value)) = map.get_next_mut(&4) {
    *value = 4;
}
assert_eq!(ordmap![1 => 1, 3 => 3, 5 => 4], map);

pub fn insert(&mut self, key: K, value: V) -> Option<V>

Insert a key/value mapping into a map.

This is a copy-on-write operation, so that the parts of the map’s structure which are shared with other maps will be safely copied before mutating.

If the map already has a mapping for the given key, the previous value is overwritten.

Time: O(log n)

Examples

let mut map = ordmap!{};
map.insert(123, "123");
map.insert(456, "456");
assert_eq!(
  map,
  ordmap!{123 => "123", 456 => "456"}
);

pub fn remove<BK>(&mut self, k: &BK) -> Option<V>
where BK: Ord + ?Sized, K: Borrow<BK>,

Remove a key/value mapping from a map if it exists.

Time: O(log n)

Examples

let mut map = ordmap!{123 => "123", 456 => "456"};
map.remove(&123);
map.remove(&456);
assert!(map.is_empty());

pub fn remove_with_key<BK>(&mut self, k: &BK) -> Option<(K, V)>
where BK: Ord + ?Sized, K: Borrow<BK>,

Remove a key/value pair from a map, if it exists, and return the removed key and value.

Time: O(log n)

pub fn update(&self, key: K, value: V) -> GenericOrdMap<K, V, P>

Construct a new map by inserting a key/value mapping into a map.

If the map already has a mapping for the given key, the previous value is overwritten.

Time: O(log n)

Examples

let map = ordmap!{};
assert_eq!(
  map.update(123, "123"),
  ordmap!{123 => "123"}
);

pub fn update_with<F>(self, k: K, v: V, f: F) -> GenericOrdMap<K, V, P>
where F: FnOnce(V, V) -> V,

Construct a new map by inserting a key/value mapping into a map.

If the map already has a mapping for the given key, we call the provided function with the old value and the new value, and insert the result as the new value.

Time: O(log n)

pub fn update_with_key<F>(self, k: K, v: V, f: F) -> GenericOrdMap<K, V, P>
where F: FnOnce(&K, V, V) -> V,

Construct a new map by inserting a key/value mapping into a map.

If the map already has a mapping for the given key, we call the provided function with the key, the old value and the new value, and insert the result as the new value.

Time: O(log n)

pub fn update_lookup_with_key<F>( self, k: K, v: V, f: F, ) -> (Option<V>, GenericOrdMap<K, V, P>)
where F: FnOnce(&K, &V, V) -> V,

Construct a new map by inserting a key/value mapping into a map, returning the old value for the key as well as the new map.

If the map already has a mapping for the given key, we call the provided function with the key, the old value and the new value, and insert the result as the new value.

Time: O(log n)

pub fn alter<F>(&self, f: F, k: K) -> GenericOrdMap<K, V, P>
where F: FnOnce(Option<V>) -> Option<V>,

Update the value for a given key by calling a function with the current value and overwriting it with the function’s return value.

The function gets an Option<V> and returns the same, so that it can decide to delete a mapping instead of updating the value, and decide what to do if the key isn’t in the map.

Time: O(log n)

pub fn without<BK>(&self, k: &BK) -> GenericOrdMap<K, V, P>
where BK: Ord + ?Sized, K: Borrow<BK>,

Remove a key/value pair from a map, if it exists.

Time: O(log n)

pub fn extract<BK>(&self, k: &BK) -> Option<(V, GenericOrdMap<K, V, P>)>
where BK: Ord + ?Sized, K: Borrow<BK>,

Remove a key/value pair from a map, if it exists, and return the removed value as well as the updated list.

Time: O(log n)

pub fn extract_with_key<BK>( &self, k: &BK, ) -> Option<(K, V, GenericOrdMap<K, V, P>)>
where BK: Ord + ?Sized, K: Borrow<BK>,

Remove a key/value pair from a map, if it exists, and return the removed key and value as well as the updated list.

Time: O(log n)

pub fn union(self, other: GenericOrdMap<K, V, P>) -> GenericOrdMap<K, V, P>

Construct the union of two maps, keeping the values in the current map when keys exist in both maps.

Time: O(n log n)

Examples

let map1 = ordmap!{1 => 1, 3 => 3};
let map2 = ordmap!{2 => 2, 3 => 4};
let expected = ordmap!{1 => 1, 2 => 2, 3 => 3};
assert_eq!(expected, map1.union(map2));

pub fn union_with<F>( self, other: GenericOrdMap<K, V, P>, f: F, ) -> GenericOrdMap<K, V, P>
where F: FnMut(V, V) -> V,

Construct the union of two maps, using a function to decide what to do with the value when a key is in both maps.

The function is called when a value exists in both maps, and receives the value from the current map as its first argument, and the value from the other map as the second. It should return the value to be inserted in the resulting map.

Time: O(n log n)

pub fn union_with_key<F>( self, other: GenericOrdMap<K, V, P>, f: F, ) -> GenericOrdMap<K, V, P>
where F: FnMut(&K, V, V) -> V,

Construct the union of two maps, using a function to decide what to do with the value when a key is in both maps.

The function is called when a value exists in both maps, and receives a reference to the key as its first argument, the value from the current map as the second argument, and the value from the other map as the third argument. It should return the value to be inserted in the resulting map.

Time: O(n log n)

Examples

let map1 = ordmap!{1 => 1, 3 => 4};
let map2 = ordmap!{2 => 2, 3 => 5};
let expected = ordmap!{1 => 1, 2 => 2, 3 => 9};
assert_eq!(expected, map1.union_with_key(
    map2,
    |key, left, right| left + right
));

pub fn unions<I>(i: I) -> GenericOrdMap<K, V, P>
where I: IntoIterator<Item = GenericOrdMap<K, V, P>>,

Construct the union of a sequence of maps, selecting the value of the leftmost when a key appears in more than one map.

Time: O(n log n)

Examples

let map1 = ordmap!{1 => 1, 3 => 3};
let map2 = ordmap!{2 => 2};
let expected = ordmap!{1 => 1, 2 => 2, 3 => 3};
assert_eq!(expected, OrdMap::unions(vec![map1, map2]));

pub fn unions_with<I, F>(i: I, f: F) -> GenericOrdMap<K, V, P>
where I: IntoIterator<Item = GenericOrdMap<K, V, P>>, F: Fn(V, V) -> V,

Construct the union of a sequence of maps, using a function to decide what to do with the value when a key is in more than one map.

The function is called when a value exists in multiple maps, and receives the value from the current map as its first argument, and the value from the next map as the second. It should return the value to be inserted in the resulting map.

Time: O(n log n)

pub fn unions_with_key<I, F>(i: I, f: F) -> GenericOrdMap<K, V, P>
where I: IntoIterator<Item = GenericOrdMap<K, V, P>>, F: Fn(&K, V, V) -> V,

Construct the union of a sequence of maps, using a function to decide what to do with the value when a key is in more than one map.

The function is called when a value exists in multiple maps, and receives a reference to the key as its first argument, the value from the current map as the second argument, and the value from the next map as the third argument. It should return the value to be inserted in the resulting map.

Time: O(n log n)

pub fn difference(self, other: GenericOrdMap<K, V, P>) -> GenericOrdMap<K, V, P>

👎Deprecated since 2.0.1: to avoid conflicting behaviors between std and imbl, the difference alias for symmetric_difference will be removed.

Construct the symmetric difference between two maps by discarding keys which occur in both maps.

This is an alias for the symmetric_difference method.

Time: O(n log n)

Examples

let map1 = ordmap!{1 => 1, 3 => 4};
let map2 = ordmap!{2 => 2, 3 => 5};
let expected = ordmap!{1 => 1, 2 => 2};
assert_eq!(expected, map1.difference(map2));

pub fn symmetric_difference( self, other: GenericOrdMap<K, V, P>, ) -> GenericOrdMap<K, V, P>

Construct the symmetric difference between two maps by discarding keys which occur in both maps.

Time: O(n log n)

Examples

let map1 = ordmap!{1 => 1, 3 => 4};
let map2 = ordmap!{2 => 2, 3 => 5};
let expected = ordmap!{1 => 1, 2 => 2};
assert_eq!(expected, map1.symmetric_difference(map2));

pub fn difference_with<F>( self, other: GenericOrdMap<K, V, P>, f: F, ) -> GenericOrdMap<K, V, P>
where F: FnMut(V, V) -> Option<V>,

👎Deprecated since 2.0.1: to avoid conflicting behaviors between std and imbl, the difference_with alias for symmetric_difference_with will be removed.

Construct the symmetric difference between two maps by using a function to decide what to do if a key occurs in both.

This is an alias for the symmetric_difference_with method.

Time: O(n log n)

pub fn symmetric_difference_with<F>( self, other: GenericOrdMap<K, V, P>, f: F, ) -> GenericOrdMap<K, V, P>
where F: FnMut(V, V) -> Option<V>,

Construct the symmetric difference between two maps by using a function to decide what to do if a key occurs in both.

Time: O(n log n)

pub fn difference_with_key<F>( self, other: GenericOrdMap<K, V, P>, f: F, ) -> GenericOrdMap<K, V, P>
where F: FnMut(&K, V, V) -> Option<V>,

👎Deprecated since 2.0.1: to avoid conflicting behaviors between std and imbl, the difference_with_key alias for symmetric_difference_with_key will be removed.

Construct the symmetric difference between two maps by using a function to decide what to do if a key occurs in both. The function receives the key as well as both values.

This is an alias for the symmetric_difference_with_key method.

Time: O(n log n)

Examples

let map1 = ordmap!{1 => 1, 3 => 4};
let map2 = ordmap!{2 => 2, 3 => 5};
let expected = ordmap!{1 => 1, 2 => 2, 3 => 9};
assert_eq!(expected, map1.difference_with_key(
    map2,
    |key, left, right| Some(left + right)
));

pub fn symmetric_difference_with_key<F>( self, other: GenericOrdMap<K, V, P>, f: F, ) -> GenericOrdMap<K, V, P>
where F: FnMut(&K, V, V) -> Option<V>,

Construct the symmetric difference between two maps by using a function to decide what to do if a key occurs in both. The function receives the key as well as both values.

Time: O(n log n)

Examples

let map1 = ordmap!{1 => 1, 3 => 4};
let map2 = ordmap!{2 => 2, 3 => 5};
let expected = ordmap!{1 => 1, 2 => 2, 3 => 9};
assert_eq!(expected, map1.symmetric_difference_with_key(
    map2,
    |key, left, right| Some(left + right)
));

pub fn relative_complement( self, other: GenericOrdMap<K, V, P>, ) -> GenericOrdMap<K, V, P>

Construct the relative complement between two maps by discarding keys which occur in other.

Time: O(m log n) where m is the size of the other map

Examples

let map1 = ordmap!{1 => 1, 3 => 4};
let map2 = ordmap!{2 => 2, 3 => 5};
let expected = ordmap!{1 => 1};
assert_eq!(expected, map1.relative_complement(map2));

pub fn intersection( self, other: GenericOrdMap<K, V, P>, ) -> GenericOrdMap<K, V, P>

Construct the intersection of two maps, keeping the values from the current map.

Time: O(n log n)

Examples

let map1 = ordmap!{1 => 1, 2 => 2};
let map2 = ordmap!{2 => 3, 3 => 4};
let expected = ordmap!{2 => 2};
assert_eq!(expected, map1.intersection(map2));

pub fn intersection_with<B, C, F, P2, P3>( self, other: GenericOrdMap<K, B, P2>, f: F, ) -> GenericOrdMap<K, C, P3>
where B: Clone, C: Clone, F: FnMut(V, B) -> C, P2: SharedPointerKind, P3: SharedPointerKind,

Construct the intersection of two maps, calling a function with both values for each key and using the result as the value for the key.

Time: O(n log n)

pub fn intersection_with_key<B, C, F, P2, P3>( self, other: GenericOrdMap<K, B, P2>, f: F, ) -> GenericOrdMap<K, C, P3>
where B: Clone, C: Clone, F: FnMut(&K, V, B) -> C, P2: SharedPointerKind, P3: SharedPointerKind,

Construct the intersection of two maps, calling a function with the key and both values for each key and using the result as the value for the key.

Time: O(n log n)

Examples

let map1 = ordmap!{1 => 1, 2 => 2};
let map2 = ordmap!{2 => 3, 3 => 4};
let expected = ordmap!{2 => 5};
assert_eq!(expected, map1.intersection_with_key(
    map2,
    |key, left, right| left + right
));

pub fn split<BK>( &self, split: &BK, ) -> (GenericOrdMap<K, V, P>, GenericOrdMap<K, V, P>)
where BK: Ord + ?Sized, K: Borrow<BK>,

Split a map into two, with the left hand map containing keys which are smaller than split, and the right hand map containing keys which are larger than split.

The split mapping is discarded.

pub fn split_lookup<BK>( &self, split: &BK, ) -> (GenericOrdMap<K, V, P>, Option<V>, GenericOrdMap<K, V, P>)
where BK: Ord + ?Sized, K: Borrow<BK>,

Split a map into two, with the left hand map containing keys which are smaller than split, and the right hand map containing keys which are larger than split.

Returns both the two maps and the value of split.

pub fn take(&self, n: usize) -> GenericOrdMap<K, V, P>

Construct a map with only the n smallest keys from a given map.

pub fn skip(&self, n: usize) -> GenericOrdMap<K, V, P>

Construct a map with the n smallest keys removed from a given map.

pub fn without_min(&self) -> (Option<V>, GenericOrdMap<K, V, P>)

Remove the smallest key from a map, and return its value as well as the updated map.

pub fn without_min_with_key(&self) -> (Option<(K, V)>, GenericOrdMap<K, V, P>)

Remove the smallest key from a map, and return that key, its value as well as the updated map.

pub fn without_max(&self) -> (Option<V>, GenericOrdMap<K, V, P>)

Remove the largest key from a map, and return its value as well as the updated map.

pub fn without_max_with_key(&self) -> (Option<(K, V)>, GenericOrdMap<K, V, P>)

Remove the largest key from a map, and return that key, its value as well as the updated map.

pub fn entry(&mut self, key: K) -> Entry<'_, K, V, P>

Get the Entry for a key in the map for in-place manipulation.

Time: O(log n)